+++ android/sdk/src/main/cpp/app/organicmaps/sdk/bookmarks/data/BookmarkInfo.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <jni.h>
+
+class Bookmark;
+
+jobject CreateBookmarkInfo(JNIEnv * env, Bookmark const & bookmark);
