--- android/sdk/src/main/cpp/app/organicmaps/sdk/bookmarks/data/BookmarkJniHelpers.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include <jni.h>
-
-class Bookmark;
-class Track;
-
-namespace bookmark_jni
-{
-void PrepareClassRefs(JNIEnv * env);
-jobject CreateBookmarkInfo(JNIEnv * env, Bookmark const & bookmark);
-jobject CreateTrack(JNIEnv * env, Track const & track);
-}  // namespace bookmark_jni
